requirements.txt
inflect==0.2.5
mysqlclient==1.3.7
py-moneyed==0.6.0
sqlacodegen==1.1.6
SQLAlchemy==1.0.14
